pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payments-service"
version = "0.1.0"
description = "Async payment processing microservice with Outbox pattern"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "faststream[rabbit]>=0.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "testcontainers[postgres,rabbitmq]>=4.7",
    "asgi-lifespan>=2.1",
    "ruff>=0.6",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "G",    # flake8-logging-format
    "LOG",  # flake8-logging
    "RUF",  # ruff-specific
]
ignore = [
    "B008",  # do not perform function call in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"migrations/**" = ["E501", "N999"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
packages = ["app"]
strict = true
plugins = ["pydantic.mypy"]
disallow_untyped_decorators = false  # FastAPI/FastStream decorators are untyped
warn_unused_ignores = true
warn_return_any = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["faststream.*", "aio_pika.*", "aiormq.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
